+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # Changelog
 Notable changes to this project.
 
+## dev
+- make `get_current_round` available in SignalsAPI
+
 ## [2.4.3] - 2021-02-27
 - deprecate multi-tournament handling
 - `get_models` now returns the list of models depending on the tournament you
   are working on - numerai classic (NumerAPI) vs numerai signals (SignalsAPI).
-  This is necessary after the recent "model split". 
+  This is necessary after the recent "model split".
 
 ## [2.4.2] - 2021-02-25
 - `daily_user_performances` add `fnc`